requirements.txt
stockfish>= 3.22.1
pytest>=7.0
pyserial>=7.0